requirements/testing.txt
coverage==4.2
defusedxml==0.4.1
Django==1.10
django-allauth==0.26.1
django-filter==0.14.0
django-material==0.8.0
django-widget-tweaks==1.4.1
djangorestframework==3.4.6
drf-nested-routers==0.11.1
httpie==0.9.6
Markdown==2.6.6
mysqlclient==1.3.7
oauthlib==1.1.2
Pillow==3.3.0
Pygments==2.1.3
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
selenium==2.53.6
